+++ src/basepkg/gp_tools/__init__.py
@@ -1,3 +1,22 @@
 # coding=utf-8
+"""Tools for development.
 
-__author__ = 'sb'
+Logger.
+Random testing tool.
+
+Copyright 2014, Stefan Braun
+Licensed under MIT.
+"""
+
+from .log import LoggerFactory
+from .fuzzer import fuzzer, fuzz_string, FuzzExecutor
+import logging
+
+
+# Symbols available when importing with *.
+__all__ = ['LoggerFactory', 'fuzzer', 'fuzz_string', 'FuzzExecutor']
+
+
+# Configure NullHandler to prevent warning in case logging is not configured.
+# See https://docs.python.org/2/howto/logging.html#library-config
+logging.getLogger('gp_tools').addHandler(logging.NullHandler())